requirements.txt
python-telegram-bot==20.7
requests
beautifulsoup4
pytz
urllib3
python-dotenv
pandas